More variable available for emailing transcript.
Implemented more variable for emailing transcript. However, I had to re-implement the whole logic for it.
Before this commit, the program send email first then submit interview data. Now it submit data first, then send email.
The reason for doing so is that the interviewID used to be generated in lambda function, which means there is no way for me to send an email with variable that haven't been created yet. I modified the code to make interviewID to be generated in the conductinterview.js and passed to the lambda function, then DynamoDB. You may want to do "npm install uuid" to allow a new import to work (It worked for me without any install, but it's safe to do so).